refactor(map): z_index 상수화 및 위치 이동
