+++ src/constants/map/index.ts
@@ -1,7 +1,3 @@
-import { days } from '@/constants/map/Days';
-import { categories } from '@/constants/map/Categories';
-import { DAYS } from '@/constants/map/Days';
-import { CATEGORIES } from '@/constants/map/Categories';
-
-export { days, categories };
-export type { DAYS, CATEGORIES };
+export * from '@/constants/map/Days';
+export * from '@/constants/map/Categories';
+export * from '@/constants/map/Z_Index';
