--- src/pages/map/MapPage.types.ts
@@ -1,5 +1,0 @@
-export const Z_INDEX = {
-  MAP: 0,
-  MAP_OVERLAY: 100,
-  CONTENTS: 200,
-};
